requirements.txt
Flask==1.0.2
flask_graphql==2.0.0
graphene==2.1.3
graphene_mongo==0.2.6
mongoengine==0.17.0
arango_orm==0.5.7
python-arango==4.4.0